Skip pip

+++ scripts/use_pyc.py
@@ -3,10 +3,11 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 import logging
+import glob
 import os
+import platform
 import re
 import sys
-import glob
 from pathlib import Path
 import shutil
 
@@ -48,6 +49,19 @@ def main(folder, version=None):
     pyc_suffix = f'cpython-{version.replace(".", "")}.pyc'
     _LOGGER.info(f'pyc suffix: {pyc_suffix}')
     for file in glob.glob(f'{folder}/**/__pycache__/*{pyc_suffix}', recursive=True):
+        # If pip's py files are also removed, the following error will occur when use pip install.
+        # For example: /opt/az/bin/python3 -m pip install --no-cache  pgcli==3.4.1
+        #   × pip subprocess to install build dependencies did not run successfully.
+        #   │ exit code: 2
+        #   ╰─> [1 lines of output]
+        #       /opt/az/bin/python3: can't open file '/opt/az/lib/python3.10/site-packages/pip/__pip-runner__.py': [Errno 2] No such file or directory
+        #       [end of output]
+        #
+        #   note: This error originates from a subprocess, and is likely not a problem with pip.
+        # error: subprocess-exited-with-error
+        if 'site-packages/pip' in file:
+            continue
+
         # file is /opt/az/lib/python3.10/site-packages/websocket/__pycache__/_app.cpython-310.pyc
         # py_filename is _app.py
         py_filename = Path(file).name[:-len('cpython-310.pyc')] + 'py'
@@ -58,6 +72,9 @@ def main(folder, version=None):
             shutil.move(file, py_path.with_suffix('.pyc'))
 
     for f in glob.glob(f'{folder}/**/__pycache__', recursive=True):
+        # Remove pip __pycache__ folder for Windows only to save more space
+        if 'site-packages/pip' in f and not platform.system() == 'Windows':
+            continue
         shutil.rmtree(f)
 
     _LOGGER.info('Finish processing')
